requirements.txt
streamlit
pandas
fpdf
psycopg2
python-dotenv
